[testing] pytests (#22)
* [testing] pytests

* add pytest pydep

* install microgbtpy

* dev script

* fix leakage in treenode

* fix gh action

* WIP

* [travis] setup env

* clean up pytests

* [git-action] upgrade pip

* run pytests as well

* [travis] use python3

* fix travis

* fix travis

* fix travis; wheel

* add github action for valgrind

* remove valgrind from travis

* minor fixes

* split py / cpp tests

* proper setup python3

* fix py2/3 issue

* merge github actions

* fix bash scripts

* minor fix tests

* pytest ubuntu py3

* split cpp and python tests

* remove -Ofast

* fix tests

* fix python install

* remove py build_ext

* sudo py install

* add find_packages

* fix suffix name

* clean up eggs

* debug ls

* remove python 2.7

* remove py2.7 as sudo

* update python deps

* install libpython3.6-dev

* apt upgrade

* ACEEPT EULA

* add clang-format

* apply clang-format

* clang-format tests/ as well

* run from root dir

* use python3 -m pytest

* install via py setup.py

* install pypackage

* sudo install py

* [github-actions] remove clang-formatter

* minor fix

+++ test/test_dataset.cpp
@@ -4,19 +4,19 @@
 TEST(Dataset, DefaultConstructor)
 {
 
-    size_t m = 2, n = 3;
+    long m = 2, n = 3;
     Eigen::MatrixXd A = Eigen::MatrixXd::Zero(m, n);
     microgbt::Vector y = {1.0, 2.0};
     microgbt::Dataset dataset(A, y);
 
-    ASSERT_EQ(dataset.nRows(),  m);
+    ASSERT_EQ(dataset.nRows(), m);
     ASSERT_EQ(dataset.numFeatures(), n);
 }
 
 TEST(Dataset, Constructor)
 {
 
-    size_t m = 2, n = 3;
+    long m = 2, n = 3;
     Eigen::MatrixXd A = Eigen::MatrixXd::Zero(m, n);
     microgbt::Vector y = {1.0, 2.0};
     microgbt::Dataset dataset(A, y);
@@ -29,5 +29,4 @@ TEST(Dataset, Constructor)
 
     ASSERT_EQ(leftDS.nRows(), left.size());
     ASSERT_EQ(leftDS.numFeatures(), n);
-
 }
